+++ packages/meteor-components-ioc-plugin/meteor-components-ioc-plugin.js
@@ -1,4 +1,4 @@
-/*global ComponentRootIoc, Package*/
+/*global ComponentRootIoc, Package, ReactiveVar*/
 
 function install(Component, ComponentUtil, IocContainer) {
   ComponentRootIoc = new IocContainer();
@@ -10,6 +10,10 @@ function install(Component, ComponentUtil, IocContainer) {
     if (!templateInstance.view.ioc) {
       let ioc = new IocContainer(nearestIoc);
       templateInstance.view.ioc = ioc;
+
+      installDataContextKeys(ioc, templateInstance);
+      installDataContext(ioc, templateInstance);
+
       ioc.install(
         componentName,
         Ctor,
@@ -54,6 +58,63 @@ function getNearestIocContainer(view, defaultIoc) {
   } while (view && !view.ioc);
 
   return view ? (view.ioc || defaultIoc) : defaultIoc;
+}
+
+function installDataContext(ioc, templateInstance) {
+  let data = templateInstance.data;
+
+  if (data) {
+    let dataVar = new ReactiveVar(data);
+    ioc.install('data', () => dataVar);
+
+    templateInstance.autorun(function (c) {
+      let data = Template.currentData();
+
+      if (!c.fristRun) {
+        dataVar.curValue = data;
+        dataVar.dep.changed();
+      }
+    });
+  }
+}
+
+function installDataContextKeys(ioc, templateInstance) {
+  let data = templateInstance.data;
+  let vars = {};
+
+  if (data) {
+    if (typeof data === 'object' && !Array.isArray(data)) {
+      for (let key in data) {
+        vars[key] = new ReactiveVar(data[key]);
+        ioc.install(key, () => vars[key]);
+      }
+    }
+
+    templateInstance.autorun(function (c) {
+      // depend on a reactive data context.
+      let data = Template.currentData();
+
+      if (c.firstRun) {
+        c.onStop(function () {
+          for (let key in vars) {
+            vars[key] = null;
+          }
+        });
+      } else {
+        // We have to forcefully mark each data context key as being
+        // changed because the new value could be the same object only
+        // with updates. We set the value manually so that we don't
+        // inadvertanly cause more than one 'change' event to occur
+        // for each data context key.
+        if (data && typeof data === 'object' && !Array.isArray(data)) {
+          for (let key in vars) {
+            vars[key].curValue = data[key];
+            vars[key].dep.changed();
+          }
+        }
+      }
+    });
+  }
 }
 
 // If the weak dependencies exist then we install the plugin.