+++ packages/meteor-components-ioc-plugin/HISTORY.md
@@ -1,3 +1,15 @@
+# 0.1.0
+
+**Dec. 30, 2015**
+
+- Install data context as service with name `data` as a `ReactiveVar` instance.
+  The instance will be marked as `changed` when the data context changes.
+
+- Install each key of the data context as a service as a `ReactiveVar`
+  instance. Each instance will be marked as `changed` when the data context
+  changes.
+
+
 # 0.0.1
 
 **Dec. 29, 2015**